Fix renaming issues
