+++ packages/api/src/beacon/client/proof.ts
